refactor : change user dto package
