+++ app/src/main/java/com/kalex/dogescollection/api/DogsApi.kt
@@ -1,14 +1,13 @@
 package com.kalex.dogescollection.api
 
 import com.kalex.dogescollection.authentication.createaccount.dto.SignUpDTO
-import com.kalex.dogescollection.authentication.createaccount.dto.UserResponse
+import com.kalex.dogescollection.authentication.model.dto.UserResponse
 import com.kalex.dogescollection.authentication.login.dto.LogInDTO
 import com.kalex.dogescollection.common.Constants
 import com.kalex.dogescollection.dogList.model.data.dto.Dogs
 import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.POST
-import retrofit2.http.PUT
 
 interface DogsApi {
