--- app/src/main/java/com/kalex/dogescollection/authentication/createaccount/dto/Data.kt
@@ -1,5 +1,0 @@
-package com.kalex.dogescollection.authentication.createaccount.dto
-
-data class UserAccountData(
-    val user: UserDTO
-)
