+++ app/src/main/java/com/kalex/dogescollection/authentication/login/presentation/LoginViewModel.kt
@@ -2,12 +2,11 @@ package com.kalex.dogescollection.authentication.login.presentation
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.kalex.dogescollection.authentication.createaccount.dto.User
+import com.kalex.dogescollection.authentication.model.dto.User
 import com.kalex.dogescollection.authentication.model.AuthenticationUseCase
 import com.kalex.dogescollection.common.networkstates.UseCaseFlowStatus
 import com.kalex.dogescollection.common.networkstates.ViewModelNewsUiState
 import dagger.hilt.android.lifecycle.HiltViewModel
-import dagger.hilt.android.scopes.ViewModelScoped
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.collectLatest
