+++ app/src/main/java/com/kalex/dogescollection/authentication/model/AuthenticationRepository.kt
@@ -1,6 +1,6 @@
 package com.kalex.dogescollection.authentication.model
 
-import com.kalex.dogescollection.authentication.createaccount.dto.UserResponse
+import com.kalex.dogescollection.authentication.model.dto.UserResponse
 
 interface AuthenticationRepository {
     suspend fun createAccount(user: String, password: String, passwordConfirm: String): UserResponse
