+++ app/src/main/java/com/kalex/dogescollection/authentication/model/dto/Data.kt
@@ -1,0 +1,5 @@
+package com.kalex.dogescollection.authentication.model.dto
+
+data class UserAccountData(
+    val user: UserDTO
+)
